refactor(member): optimized book-loan API cache (#152)
