+++ apps/member/src/hooks/queries/book-loan/useBookLoanExtendMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { postExtendBook } from '@api/book';
-import { BOOK_LOAN_RECORD_QUERY_KEY, MY_BOOK_QUERY_KEY } from '@constants/key';
+import { BOOK_LOAN_RECORD_QUERY_KEY } from '@constants/key';
 import { API_ERROR_MESSAGE } from '@constants/message';
 import useToast from '@hooks/common/useToast';
 
@@ -12,16 +12,14 @@ export function useBookLoanExtendMutation() {
   const queryClient = useQueryClient();
   const toast = useToast();
 
-  const bookExtendMutation = useMutation({
+  const mutation = useMutation({
     mutationFn: postExtendBook,
     onSuccess: ({ success, errorMessage }, { borrowerId }) => {
       if (success) {
         queryClient.invalidateQueries({
-          queryKey: MY_BOOK_QUERY_KEY.BOOKS(),
-        });
-        queryClient.invalidateQueries({
           queryKey: BOOK_LOAN_RECORD_QUERY_KEY.BORROWER(borrowerId),
         });
+
         toast({
           state: 'success',
           message: '해당 도서 대여 기간을 연장했어요.',
@@ -35,5 +33,5 @@ export function useBookLoanExtendMutation() {
     },
   });
 
-  return { bookExtendMutate: bookExtendMutation.mutate };
+  return { bookExtendMutate: mutation.mutate };
 }
