+++ apps/member/src/hooks/queries/book-loan/useBookLoanRecordApproveMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { patchBookLoanRecordApprove } from '@api/book';
-import { BOOK_LOAN_RECORD_QUERY_KEY, MY_BOOK_QUERY_KEY } from '@constants/key';
+import { BOOK_LOAN_RECORD_QUERY_KEY, BOOK_QUERY_KEY } from '@constants/key';
 import { API_ERROR_MESSAGE } from '@constants/message';
 import useToast from '@hooks/common/useToast';
 
@@ -12,16 +12,23 @@ export function useBookLoanRecordApproveMutation() {
   const queryClient = useQueryClient();
   const toast = useToast();
 
-  const bookLoanRecordApprove = useMutation({
+  const mutation = useMutation({
     mutationFn: patchBookLoanRecordApprove,
-    onSuccess: ({ success, errorMessage }) => {
+    onSuccess: ({ success, errorMessage }, id) => {
       if (success) {
         queryClient.invalidateQueries({
-          queryKey: MY_BOOK_QUERY_KEY.BOOKS(),
+          queryKey: BOOK_QUERY_KEY.DETAIL(id),
+        });
+        queryClient.invalidateQueries({
+          queryKey: BOOK_LOAN_RECORD_QUERY_KEY.BOOK(id),
         });
         queryClient.invalidateQueries({
           queryKey: BOOK_LOAN_RECORD_QUERY_KEY.BORROWERS(),
         });
+        queryClient.invalidateQueries({
+          queryKey: BOOK_LOAN_RECORD_QUERY_KEY.RECORDS_PAGE(),
+        });
+
         toast({
           state: 'success',
           message: '해당 도서 대여를 승인했어요.',
@@ -32,5 +39,5 @@ export function useBookLoanRecordApproveMutation() {
     },
   });
 
-  return { bookLoanRecordApproveMutate: bookLoanRecordApprove.mutate };
+  return { bookLoanRecordApproveMutate: mutation.mutate };
 }
