+++ apps/member/src/hooks/queries/book-loan/useBookLoanRecordConditions.ts
@@ -5,7 +5,7 @@ import { BOOK_LOAN_RECORD_QUERY_KEY } from '@constants/key';
 
 import type { WithPaginationParams } from '@type/api';
 
-interface UseBookLoanRecordConditionsPrams extends WithPaginationParams {
+interface Prams extends WithPaginationParams {
   bookId?: number;
   borrowerId?: string;
   isReturned?: boolean;
@@ -20,12 +20,14 @@ export function useBookLoanRecordConditions({
   isReturned,
   page = 0,
   size = 20,
-}: UseBookLoanRecordConditionsPrams) {
-  const queryKey = bookId
+}: Prams) {
+  const bookQueryKey = bookId
     ? BOOK_LOAN_RECORD_QUERY_KEY.BOOK(bookId)
-    : borrowerId
-      ? BOOK_LOAN_RECORD_QUERY_KEY.BORROWER(borrowerId)
-      : BOOK_LOAN_RECORD_QUERY_KEY.ALL;
+    : undefined;
+
+  const borrowerQueryKey = borrowerId
+    ? BOOK_LOAN_RECORD_QUERY_KEY.BORROWER(borrowerId)
+    : BOOK_LOAN_RECORD_QUERY_KEY.RECORD({ page, size });
 
   return useSuspenseQuery({
     queryFn: () =>
@@ -36,6 +38,6 @@ export function useBookLoanRecordConditions({
         page,
         size,
       }),
-    queryKey: queryKey,
+    queryKey: bookQueryKey || borrowerQueryKey,
   });
 }
