+++ apps/member/src/hooks/queries/book-loan/useBookLoanRecordOverdue.ts
@@ -5,21 +5,19 @@ import { BOOK_LOAN_RECORD_QUERY_KEY } from '@constants/key';
 
 import type { WithPaginationParams } from '@type/api';
 
-interface UseBookLoanRecordOverduePrams extends WithPaginationParams {}
-
 /**
  * 도서 연체자를 조회합니다.
  */
 export function useBookLoanRecordOverdue({
   page = 0,
   size = 20,
-}: UseBookLoanRecordOverduePrams = {}) {
+}: WithPaginationParams = {}) {
   return useSuspenseQuery({
+    queryKey: BOOK_LOAN_RECORD_QUERY_KEY.OVERDUE({ page, size }),
     queryFn: () =>
       getBookLoanRecordOverdue({
         page,
         size,
       }),
-    queryKey: BOOK_LOAN_RECORD_QUERY_KEY.OVERDUE({ page, size }),
   });
 }
